BINIUS-90: make PackedField a Field supertrait
Every field is a packed vector of width one holding itself, so declare
that in the type system: `Field: PackedField<Scalar = Self>`.

The former blanket `impl<F: Field> PackedField for F` is replaced by a
per-type width-one impl emitted from the `binary_field!` macro, matching
how the macro already emits the reflexive `Divisible<Self>`/`Maskable<Self>`
impls. The impl bodies are byte-identical to the old blanket, so behavior
is unchanged.

With `Field` now inheriting the full packed API transitively, its
supertrait list collapses to the field-only additions (Display, Hash,
serialization); the arithmetic ops, Divisible/Maskable, Random, Zeroable,
and WideMul bounds are all supplied by PackedField + FieldOps.

Pure refactor, no behavior change.
